requirements.txt
streamlit==1.25.0
pandas>=2.0
Pillow>=10.0
snowflake-connector-python>=3.0.0
python-dotenv>=1.0
numpy>=1.25
